+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 - Files: cron job file notification
+- Core: updated german translation
+- Various: route to main grid when details resets / is deleted for mobile view
 
 11-11-2024: 6.8.85
 - Billing: Fixed undefined index 'sort'